backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-rest-auth>=7.2.0",
    "django==5.2.12",
    "django-allauth>=65.15.1",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-silk>=5.5.0",
    "django-summernote>=0.8.20.0",
    "djangorestframework>=3.17.1",
    "psycopg>=3.3.3",
]
[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.5.0",
]
